backend/requirements.txt
asyncpg
psycopg2-binary
sqlalchemy
fastapi
uvicorn
python-dotenv
groq
pytubefix
youtube-transcript-api